Move clean_terms into kt utils, delete text_utils module
